pyproject.toml
[project]
name = "knowledge-base"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.20",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.75",
    "langchain-google-vertexai>=2.0.28",
    "langchain-huggingface>=0.3.1",
    "langchain-ollama>=0.3.7",
    "langchain-openai>=0.3.32",
    "langchain-text-splitters>=0.3.11",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "sentence-transformers>=5.1.0",
    "unstructured>=0.18.14",
]

[project.scripts]
knowledge-base = "knowledge_base:main"

[build-system]
requires = ["uv_build>=0.8.9,<0.9.0"]
build-backend = "uv_build"
